Fix func reference in admin
